+++ Source/Honeycomber/Public/Actors/MoneyStorage.h
@@ -18,8 +18,8 @@ class HONEYCOMBER_API AMoneyStorage : public AActor
 public:	
 	// Sets default values for this actor's properties
 	AMoneyStorage();
-	void AddMoney(int32 newMoney);
-	int32 GetMoney() { return MoneyAmount; };
+	void AddMoney(uint8 newMoney);
+	uint8 GetMoney() { return MoneyAmount; };
 
 protected:
 	// Called when the game starts or when spawned
@@ -31,5 +31,5 @@ private:
 	UPROPERTY(EditAnywhere)
 		UWidgetComponent* MoneyDisplayComponent;
 	UStateDisplay* MoneyDisplay;
-	int32 MoneyAmount;
+	uint8 MoneyAmount;
 };
